pyproject.toml
[project]
name = "webhook-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.11",
    "chromadb>=1.0.15",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.3",
    "uvicorn[standard]>=0.35.0",
]